First attempt to distrubute models
